+++ config.conf
@@ -1,8 +1,8 @@
 [global]
 
 # --- Logging ---
-log.access_file = ''
-log.error_file = ''
+log.access_file = '/tmp/pynance_access.log'
+log.error_file = '/tmp/pynance_error.log'
 
 # --- Database ---
 # [mongo | tiny] For MongoDB Server or TinyDB local file
@@ -10,12 +10,10 @@ database.backend = 'tiny'
 
 # For tiny: /path/to/
 # For mongo: mongodb://mongoadmin:secret@localhost:27017
-database.uri = 'E:\TMP'
+database.uri = '/tmp'
 
 # For tiny: Filename
 # For mongo: Collection name
-database.name = 'testdata.json'
-
-[account]
+database.name = 'pynance.json'
 
 iban = 'DE38250400660997788666'
