+++ requirements.txt
@@ -1,3 +1,4 @@
 tinydb==4.7.1
 pymongo==4.3.3
 cherrypy[routes_dispatcher,doc,testing]==18.8.0
+pytest
