+++ testdata/config.conf
@@ -1,0 +1,19 @@
+[global]
+
+# --- Logging ---
+log.access_file = ''
+log.error_file = ''
+
+# --- Database ---
+# [mongo | tiny] For MongoDB Server or TinyDB local file
+database.backend = 'tiny'
+
+# For tiny: /path/to/
+# For mongo: mongodb://mongoadmin:secret@localhost:27017
+database.uri = '/tmp'
+
+# For tiny: Filename
+# For mongo: Collection name
+database.name = 'testdata.json'
+
+iban = 'DE38250400660997788666'
